fix(ci): line-limit maintainability — split oversized outdated-resolution part_001
